+++ Rebus.ServiceProvider/ServiceCollectionExtensions.Bus.cs
@@ -5,6 +5,7 @@ using Microsoft.Extensions.Hosting;
 using Rebus.Bus;
 using Rebus.Config;
 using Rebus.Pipeline;
+using Rebus.Startup;
 
 namespace Rebus.ServiceProvider
 {
@@ -14,52 +15,53 @@ namespace Rebus.ServiceProvider
         /// Registers and/or modifies Rebus configuration for the current service collection.
         /// </summary>
         /// <param name="services">The current message service builder.</param>
-        /// <param name="configureRebus">The optional configuration actions for Rebus.</param>
-        public static IServiceCollection AddRebus(this IServiceCollection services, Func<RebusConfigurer, RebusConfigurer> configureRebus)
+        /// <param name="configure">The optional configuration actions for Rebus.</param>
+        public static IServiceCollection AddRebus(this IServiceCollection services, Func<RebusConfigurer, RebusConfigurer> configure)
         {
             if (services == null) throw new ArgumentNullException(nameof(services));
-            if (configureRebus == null) throw new ArgumentNullException(nameof(configureRebus));
-            
-            return AddRebus(services, (c, p) => configureRebus(c));
+            if (configure == null) throw new ArgumentNullException(nameof(configure));
+
+            return AddRebus(services, (c, p) => configure(c));
         }
 
         /// <summary>
         /// Registers and/or modifies Rebus configuration for the current service collection.
         /// </summary>
         /// <param name="services">The current message service builder.</param>
-        /// <param name="configureRebus">The optional configuration actions for Rebus.</param>
-        public static IServiceCollection AddRebus(this IServiceCollection services, Func<RebusConfigurer, IServiceProvider, RebusConfigurer> configureRebus)
+        /// <param name="configure">The optional configuration actions for Rebus.</param>
+        public static IServiceCollection AddRebus(this IServiceCollection services, Func<RebusConfigurer, IServiceProvider, RebusConfigurer> configure)
         {
             if (services == null) throw new ArgumentNullException(nameof(services));
-            if (configureRebus == null) throw new ArgumentNullException(nameof(configureRebus));
+            if (configure == null) throw new ArgumentNullException(nameof(configure));
 
-            var messageBusRegistration = services.FirstOrDefault(descriptor => descriptor.ServiceType == typeof(IBus));
+            var busAlreadyRegistered = services.Any(descriptor => descriptor.ServiceType == typeof(IBus));
 
-            if (messageBusRegistration != null)
+            if (busAlreadyRegistered)
             {
-                throw new InvalidOperationException(@"Sorry, but it seems like Rebus has already been configured in this service collection.");
+                throw new InvalidOperationException(@"Sorry, but it seems like Rebus has already been configured in this service collection. 
+
+It is advised to use one container instance per bus instance, because this way it can be treated as an autonomous component with the container as the root.");
             }
 
             services.AddTransient(s => MessageContext.Current ?? throw new InvalidOperationException("Attempted to resolve IMessageContext outside of a Rebus handler, which is not possible. If you get this error, it's probably a sign that your service provider is being used outside of Rebus, where it's simply not possible to resolve a Rebus message context. Rebus' message context is only available to code executing inside a Rebus handler."));
-            services.AddTransient(s => s.GetService<IBus>().Advanced.SyncBus);
+            services.AddTransient(s => s.GetRequiredService<IBus>().Advanced.SyncBus);
 
             // Register the Rebus Bus instance, to be created when it is first requested.
             services.AddSingleton(provider => new DependencyInjectionHandlerActivator(provider));
             services.AddSingleton(provider =>
             {
-                var configurer = Configure.With(provider.GetRequiredService<DependencyInjectionHandlerActivator>());
+                var activator = provider.GetRequiredService<DependencyInjectionHandlerActivator>();
 
-                configureRebus(configurer, provider);
+                var configurer = Configure.With(activator);
 
-                var bus = configurer.Start();
+                configure(configurer, provider);
 
-                // if we can, we hook up to the application's lifetime events and ensure that Rebus stops, when the application stops,
-                // thus making it so that no messages are handled while the container gets disposed
-                var applicationLifetime = provider.GetService<IApplicationLifetime>();
-                applicationLifetime?.ApplicationStopping.Register(() => bus.Advanced.Workers.SetNumberOfWorkers(0));
+                var starter = configurer.Create();
 
-                return bus;
+                return starter;
             });
+            services.AddSingleton(provider => provider.GetRequiredService<IBusStarter>().Bus);
+            services.AddSingleton(provider => new ServiceCollectionBusDisposalFacility(provider.GetRequiredService<IBus>()));
 
             return services;
         }