+++ CHANGELOG.md
@@ -35,11 +35,12 @@
 
 * Make `AutoRegister(...)` methods return `IServiceCollection` for the sake of the builder pattern - thanks [weisro]
 
-## 5.0.0-b04
+## 5.0.0-b05
 
 * Remove ASP.NET Core dependency and make it into an adapter for Microsoft.Extensions.DependencyInjection (the way it's supposed to be) - thanks [Hawxy]
 * Speed up resolution by a factor of 3 to 4
-* Detect `IApplicationLifetime` in the container and stop the bus in the `ApplicationStopping` callback
+* Cleaner separation between registration (stuff that happens to the service collection) and resolution (stuff that happens to the service provider)
+* Detect `ObjectDisposedException` when resolving handlers and interpret that as we're being shut down
 
 
 [Hawxy]: https://github.com/Hawxy